// src/app/models/mtbl/mtbls/mtbls-user.interface.ts
export interface MtblsUser {
    address: string;
    affiliation: string;
    curator: boolean;
    dbPassword: string;
    email: string;
    firstName: string;
    fullName: string;
    joinDate: any;
    lastName: string;
    orcid: string;
    role: any;
    status: any;
    userId: number;
    userName: string;
    userVerifyDbPassword?: string;
    mobilePhoneNumber?: string;
    officePhoneNumber?: string;
    apiToken: string;
}